+++ workflow/scripts/chunk_chromosomes.py
@@ -18,6 +18,8 @@ from datetime import datetime
 
 warnings.simplefilter('ignore')
 
+from utils import bitmask_to_arrays
+from utils import ratemap_to_text
 
 # --- lib --- #
 
@@ -50,7 +52,7 @@ def write_minimal_vcf(handle, sample_names, CHROM, POS, ID, REF, ALT, GT):
 
 logfile = open(snakemake.log.log, "w")
 logfile.write(f"{tag()} Random seed {snakemake.params.seed}\n")
-np.random.seed(snakemake.params.seed)
+rng = np.random.default_rng(snakemake.params.seed)
 stratify = snakemake.params.stratify
 
 vcf_file = snakemake.input.vcf
@@ -81,7 +83,7 @@ if not os.path.exists(mask_file):
     logfile.write(f"{tag()} Did not find {mask_file}, assuming the entire sequence is accessible\n")
     bedmask = np.empty((0, 2))
 else:
-    bedmask = np.loadtxt(mask_file, usecols=[1, 2]).astype(np.int64)
+    bedmask = np.loadtxt(mask_file, usecols=[1, 2], ndmin=2).astype(np.int64)
     assert np.max(bedmask) <= hapmap.sequence_length, "Mask position exceeds hapmap length"
     logfile.write(f"{tag()} Read {bedmask.shape[0]} inaccessible intervals from {mask_file}\n")
 bitmask = np.full(int(hapmap.sequence_length), False)
@@ -145,7 +147,8 @@ assert np.max(positions) <= hapmap.sequence_length, "VCF position exceeds hapmap
 # `adjusted_mu = mu * (filtered_intervals / sequence_length) * (filtered_sites / segregating_sites)`
 #
 # where the last term is calculated *discounting* sites that lie in filtered
-# intervals.
+# intervals. Alternatively, we can set the mutation rate to zero within each 
+# filtered interval, and omit the corresponding term from the adjusted rate.
 
 
 # filter variants
@@ -196,17 +199,41 @@ all_positions = positions.copy()
 genotypes, positions = genotypes[retain], positions[retain]
 
 
-# divvy chromosome into chunks
-chunk_size = snakemake.params.chunk_size
+# find inaccessible intervals greater than a certain size, and use these
+# to delimit the chunks where SINGER will be run. If chunks are larger than
+# a certain size, split them into equal subdivisions.
 mutation_rate = snakemake.params.mutation_rate
+max_chunk_size = snakemake.params.chunk_size
+sequence_length = bitmask.size
+min_gap_size = snakemake.params.gap_size
 lower = np.min(positions) - 1 
 upper = np.max(positions) + 1 
-windows = np.linspace(lower, upper, int((upper - lower) / chunk_size) + 1)
-windows = np.unique(np.append(np.append(0, windows), hapmap.sequence_length))
-windows = windows.astype(np.int64)
+gap, intervals = bitmask_to_arrays(bitmask)
+gap = np.logical_and(np.diff(intervals) > min_gap_size, gap > 0)
+intervals = intervals.astype(np.int64)
+gapmask = np.full(bitmask.size, False)
+gapmask[:int(lower)] = True
+gapmask[int(upper):] = True
+for (a, b) in zip(intervals[:-1][gap], intervals[1:][gap]):
+    gapmask[a:b] = True
+_, windows = bitmask_to_arrays(~gapmask)
 logfile.write(
-    f"{tag()} Chunking chromosome into {windows.size - 1} windows "
-    f"from {windows[0]} to {windows[-1]}\n"
+    f"{tag()} Found {gap.sum()} masked intervals larger than {min_gap_size} bp "
+    f"(including flanks outside of variant positions) that delimit {windows.size - 1} chunks\n"
+)
+windows_refined = [0.0]
+for (a, b) in zip(windows[:-1], windows[1:]):
+    assert a == windows_refined[-1]
+    if b - a > max_chunk_size:
+        subdivisions = int(np.ceil((b - a) / max_chunk_size))
+        windows_refined.extend(np.linspace(a, b, subdivisions + 1)[1:])
+    else:
+        windows_refined.append(b)
+assert windows_refined[-1] == sequence_length
+windows = np.array(windows_refined).astype(np.int64)
+logfile.write(
+    f"{tag()} Split large chunks to maximum size {max_chunk_size}, "
+    f"resulting in {windows.size - 1} chunks\n"
 )
 
 
@@ -234,14 +261,15 @@ logfile.write(f"{tag()} Counted {sum(num_retained)} retained and {sum(num_filter
 
 
 # filter chunks with too much missingness or zero recombination rate
-prop_inaccessible = (num_bases - num_accessible) / num_bases  #<< prop_missing
-prop_segregating = num_retained / num_accessible  #<< prop_snp
+prop_inaccessible = (num_bases - num_accessible) / num_bases
+prop_segregating = num_retained / num_accessible
 prop_segregating[np.isnan(prop_segregating)] = 0.0
-prop_filtered = num_filtered / (num_retained + num_filtered)  #<<<new
+prop_filtered = num_filtered / (num_retained + num_filtered)
 prop_filtered[np.isnan(prop_filtered)] = 1.0
 filter_chunks = np.logical_and.reduce([
     prop_inaccessible < snakemake.params.max_missing, 
     prop_filtered < snakemake.params.max_missing,
+    num_retained > 2,  # SINGER assertion
     prop_segregating > 0.0,
     rec_rate > 0.0,
 ])
@@ -304,6 +332,7 @@ assert num_bases.sum() == bitmask.size
 assert num_accessible.sum() == np.sum(~bitmask)
 assert num_filtered.sum() == filtered_positions.size
 assert num_retained.sum() == positions.size == np.sum(retain)
+# NB: these *may* be recalculated at finer scale to adjust mutation rate
 
 
 # calculate windowed stats and get ballpark Ne estimate from global diversity
@@ -321,8 +350,10 @@ tajima_d, *_ = allel.windowed_tajima_d(
     counts,
     windows=np.column_stack([windows[:-1] + 1, windows[1:]]),
 )
-folded_afs = allel.sfs_folded(counts, n=2 * samples.size) / np.sum(~bitmask)
-unfolded_afs = allel.sfs(counts[:, 1], n=2 * samples.size) / np.sum(~bitmask)
+if snakemake.params.polarised:
+    afs = allel.sfs(counts[:, 1], n=2 * samples.size) / np.sum(~bitmask) 
+else:
+    afs = allel.sfs_folded(counts, n=2 * samples.size) / np.sum(~bitmask)
 adjustment = sum(num_retained) / sum(num_retained + num_filtered)
 Ne = (
     allel.sequence_diversity(positions, counts, is_accessible=~bitmask) *
@@ -334,52 +365,111 @@ logfile.write(
 )
 
 
-# adjust mutation rate to account for missing data in each chunk
+# adjust mutation rate to account for missing data in each chunk,
+# either by making a rate map where missing sequence has zero rate
+# or incorporating missing sequence into the global mutation rate
+if snakemake.params.model_missing:  
+    logfile.write(
+        f"{tag()} Adjusting per-chunk mutation rate to account for filtered sites "
+        f"and setting mutation rate to zero within masked intervals\n"
+    )
+    # replace `prop_inaccessible` with a per-base binary mask
+    prop_inaccessible, breakpoints = bitmask_to_arrays(bitmask, insert_breakpoints=windows)
+    assert breakpoints[-1] == windows[-1] and breakpoints[0] == windows[0]
+    # map `prop_filtered` onto fine-scale intervals (e.g. use average within chunk)
+    # FIXME: would it be better to have a fine-scale `prop_filtered`? What about ones?
+    window_index = np.digitize(breakpoints[:-1], windows) - 1
+    assert window_index.max() == windows.size - 2 and window_index.min() == 0
+    prop_filtered = prop_filtered[window_index]
+    logfile.write(
+        f"{tag()} Fine-scale mutation rate map has {breakpoints.size - 1} "
+        f"intervals, with {int(prop_inaccessible.sum())} that are masked\n"
+    )
+else:
+    breakpoints = windows.copy()
+    logfile.write(
+        f"{tag()} Adjusting per-chunk mutation rate to account for filtered sites "
+        f"and masked intervals\n"
+    )
+adj_mu = (1 - prop_inaccessible) * (1 - prop_filtered) * mutation_rate
+
+# dump recombination rates, adjusted mutation rates, proportion inaccessible,
+# proportion filtered, and chunk boundaries as RateMaps
+adjusted_mu = msprime.RateMap(
+    position=np.array(breakpoints),
+    rate=np.array(adj_mu),
+)
+inaccessible = msprime.RateMap(
+    position=np.array(breakpoints),
+    rate=np.array(prop_inaccessible),
+)
+filtered = msprime.RateMap(
+    position=np.array(breakpoints), 
+    rate=np.array(prop_filtered),
+)
+chunks = msprime.RateMap(
+    position=np.array(windows), 
+    rate=filter_chunks.astype(float),
+)
+pickle.dump(hapmap, open(snakemake.output.recomb_rate, "wb"))
+pickle.dump(adjusted_mu, open(snakemake.output.mut_rate, "wb"))
+pickle.dump(inaccessible, open(snakemake.output.inaccessible, "wb"))
+pickle.dump(filtered, open(snakemake.output.filtered, "wb"))
+pickle.dump(chunks, open(snakemake.output.windows, "wb"))
+open(snakemake.output.mut_map, "w").write(ratemap_to_text(adjusted_mu))
+
+
+# dump SINGER parameters for each chunk
 chunks_dir = snakemake.output.chunks
 os.makedirs(f"{chunks_dir}")
-seeds = np.random.randint(0, 2**10, size=filter_chunks.size)
+seeds = rng.integers(0, 2 ** 10, size=(filter_chunks.size, 2))
 vcf_prefix = snakemake.output.vcf.removesuffix(".vcf")
-adj_mu = np.zeros(windows.size - 1)
 for i in np.flatnonzero(filter_chunks):
     start, end = windows[i], windows[i + 1]
-    adj_mu[i] = (1 - prop_inaccessible[i]) * (1 - prop_filtered[i]) * mutation_rate
     polar = 0.99 if snakemake.params.polarised else 0.5
     id = f"{i:>06}"
-    chunk_params = {
+    # SINGER truncates coordinates of SNPs, so adjust maps accordingly
+    mut = adjusted_mu.slice(start, end, trim=True)
+    rec = hapmap.slice(start, end, trim=True)
+    chunk_path = os.path.join(chunks_dir, id)
+    mutation_map_path = f"{chunk_path}.mut"
+    recombination_map_path = f"{chunk_path}.rec"
+    singer_params = {
         "thin": int(snakemake.params.mcmc_thin), 
         "n": int(snakemake.params.mcmc_samples),
         "Ne": float(Ne),
-        "m": float(adj_mu[i]), 
+        #"m": float(mut.mean_rate),
+        #"r": float(rec.mean_rate),
+        "m": str(0.0),    # NB: SINGER expects these arguments and toggles
+        "r": float(0.0),  # map usage based on if at least one is positive
+        "mut_map": str(mutation_map_path),
+        "recomb_map": str(recombination_map_path),
         "input": str(vcf_prefix), 
         "start": int(start), 
         "end": int(end), 
         "polar": float(polar),
-        "r": float(rec_rate[i]), 
-        "seed": int(seeds[i]),
-        "output": str(f"{chunks_dir}/{id}"),
+        "seed": int(seeds[i, 0]),
+        "output": str(chunk_path),
     }
-    chunk_path = f"{chunks_dir}/{id}.yaml"
-    yaml.dump(chunk_params, open(chunk_path, "w"), default_flow_style=False)
+    polegon_params = {
+        "Ne": float(Ne),
+        "mutation_map": str(mutation_map_path),
+        "num_samples": int(snakemake.params.polegon_mcmc_samples),
+        "burn_in": int(snakemake.params.polegon_mcmc_burnin),
+        "thin": int(snakemake.params.polegon_mcmc_thin),
+        "scaling_rep": int(snakemake.params.polegon_scaling_reps),
+        "max_step": float(snakemake.params.polegon_max_step),
+        "seed": int(seeds[i, 1]),
+    }
+    chunk_params = {"singer": singer_params, "polegon": polegon_params}
+    yaml.dump(
+        chunk_params, 
+        open(f"{chunk_path}.yaml", "w"), 
+        default_flow_style=False,
+    )
+    open(mutation_map_path, "w").write(ratemap_to_text(mut))
+    open(recombination_map_path, "w").write(ratemap_to_text(rec))
     logfile.write(f"{tag()} Parameters for chunk {id} are {chunk_params}\n")
-
-
-# dump adjusted mutation rates, proportion accessible, and chunk coordinates
-adjusted_mu = msprime.RateMap(
-    position=windows, 
-    rate=np.array(adj_mu),
-)
-pickle.dump(adjusted_mu, open(snakemake.output.adjusted_mu, "wb"))
-inaccessible = msprime.RateMap(
-    position=windows, 
-    rate=np.array(prop_inaccessible),
-)
-pickle.dump(inaccessible, open(snakemake.output.inaccessible, "wb"))
-filtered = msprime.RateMap(
-    position=windows, 
-    rate=np.array(prop_filtered),
-)
-pickle.dump(filtered, open(snakemake.output.filtered, "wb"))
-
 
 
 # dump filtered vcf
@@ -401,8 +491,7 @@ tajima_d[~filter_chunks] = np.nan
 vcf_stats = {
     "diversity": diversity, 
     "tajima_d": tajima_d, 
-    "folded_afs": folded_afs, 
-    "unfolded_afs": unfolded_afs,
+    "afs": afs, 
 }
 pickle.dump(vcf_stats, open(snakemake.output.vcf_stats, "wb"))
 pickle.dump(metadata, open(snakemake.output.metadata, "wb"))
@@ -419,8 +508,7 @@ if stratify is not None:
     strata_counts = genotypes.count_alleles_subpops(sample_sets, max_allele=1)
 
     strata_divergence = []
-    strata_folded_afs = []
-    strata_unfolded_afs = []
+    strata_afs = []
     for i in range(len(strata)):
         for j in range(i, len(strata)):
             divergence, *_ = allel.windowed_divergence(
@@ -433,23 +521,22 @@ if stratify is not None:
             )
             strata_divergence.append(divergence)
 
-        folded_afs = allel.sfs_folded(
-            strata_counts[strata[i]], 
-            n=2 * len(sample_sets[strata[i]]),
-        ) / np.sum(~bitmask)
-        strata_folded_afs.append(folded_afs)
-
-        unfolded_afs = allel.sfs(
-            strata_counts[strata[i]][:, 1], 
-            n=2 * len(sample_sets[strata[i]]),
-        ) / np.sum(~bitmask)
-        strata_unfolded_afs.append(unfolded_afs)
+        if snakemake.params.polarised:
+            afs = allel.sfs(
+                strata_counts[strata[i]][:, 1], 
+                n=2 * len(sample_sets[strata[i]]),
+            ) / np.sum(~bitmask)
+        else:
+            afs = allel.sfs_folded(
+                strata_counts[strata[i]], 
+                n=2 * len(sample_sets[strata[i]]),
+            ) / np.sum(~bitmask)
+        strata_afs.append(afs)
 
     strata_divergence = np.stack(strata_divergence).T
     vcf_strata_stats = {
         "strata": strata,
         "divergence": strata_divergence,
-        "folded_afs": strata_folded_afs,
-        "unfolded_afs": strata_unfolded_afs,
+        "afs": strata_afs,
     }
 pickle.dump(vcf_strata_stats, open(snakemake.output.vcf_strata_stats, "wb"))