+++ workflow/scripts/tree_statistics.py
@@ -1,5 +1,6 @@
 """
-Calculate some branch statistics from a tree sequence.
+Calculate branch statistics from a tree sequence.  This is done via simulation,
+so as to account for missing data without excessive windowing.
 
 Part of https://github.com/nspope/singer-snakemake.
 """
@@ -23,45 +24,53 @@ def tag():
 
 # --- implm --- #
 
-inaccessible = pickle.load(open(snakemake.input.inaccessible, "rb"))
-filtered = pickle.load(open(snakemake.input.filtered, "rb"))
-mutation_rate = snakemake.params.mutation_rate
-chunk_size = np.diff(inaccessible.position)
-accessible_size = chunk_size * (1 - inaccessible.rate)
-ts = tszip.decompress(snakemake.input.trees)
+windows = pickle.load(open(snakemake.input.windows, "rb"))
+adjusted_mu = pickle.load(open(snakemake.input.mut_rate, "rb"))
+seed = 1 + int(snakemake.wildcards.rep) * 1000
 
-diversity = \
-    ts.diversity(mode='branch', windows=inaccessible.position, span_normalise=True) * \
-    (1 - filtered.rate) * mutation_rate
-diversity[inaccessible.rate == 1.0] = np.nan
+accessible = msprime.RateMap(
+    position=adjusted_mu.position,
+    rate=adjusted_mu.rate > 0,
+)
+accessible_bp = np.diff(accessible.get_cumulative_mass(windows.position))
 
-# NB: these are per-chunk; correctly calculating at the sequence level
-# would require calculating numerator and denominator separately, correcting,
-# and summing across chunks
-tajima_d = ts.Tajimas_D(mode='branch', windows=inaccessible.position)  # correction cancels within chunk
-tajima_d[inaccessible.rate == 1.0] = np.nan
+ts = msprime.sim_mutations(
+    tszip.decompress(snakemake.input.trees), 
+    rate=adjusted_mu, 
+    random_seed=seed, 
+    keep=False,
+)
 
-# NB: "global" statistics have to be weighted by accessible sequence length per chunk
-correction = (1 - filtered.rate) * accessible_size / np.sum(accessible_size)
+diversity = ts.diversity(
+    mode='site', 
+    windows=windows.position, 
+    span_normalise=False,
+)
+diversity[windows.rate == 1.0] /= accessible_bp[windows.rate == 1.0]
+diversity[windows.rate == 0.0] = np.nan
 
-folded_afs = \
-    ts.allele_frequency_spectrum(mode='branch', windows=inaccessible.position, span_normalise=True)
-folded_afs = np.sum(folded_afs * correction[:, np.newaxis] * mutation_rate, axis=0)
+tajima_d = ts.Tajimas_D(
+    mode='site', 
+    windows=windows.position, 
+)
+tajima_d[windows.rate == 0.0] = np.nan
 
-unfolded_afs = \
-    ts.allele_frequency_spectrum(mode='branch', windows=inaccessible.position, span_normalise=True, polarised=True) 
-unfolded_afs = np.sum(unfolded_afs * correction[:, np.newaxis] * mutation_rate, axis=0)
+afs = ts.allele_frequency_spectrum(
+    mode='site', 
+    span_normalise=False,
+    polarised=snakemake.params.polarised,
+) / accessible_bp[windows.rate == 1.0].sum()
 
 stats = {
     "diversity": diversity, 
     "tajima_d": tajima_d, 
-    "folded_afs": folded_afs, 
-    "unfolded_afs": unfolded_afs,
+    "afs": afs, 
 }
 pickle.dump(stats, open(snakemake.output.stats, "wb"))
 
+# TODO: add proportion multimapped sites, to use as a trace diagnostic
 
-# TODO: clean this up
+# TODO: populations are now encoded directly, so use ts.samples(population=i)
 # stratified summary stats
 strata_stats = {}
 if snakemake.params.stratify is not None:
@@ -71,33 +80,29 @@ if snakemake.params.stratify is not None:
         sample_sets[strata].extend(ind.nodes)
     strata = [n for n in sample_sets.keys()]
     strata_stats["strata"] = strata
+    indexes = list(itertools.combinations_with_replacement(range(len(strata)), 2))
 
-    strata_divergence = \
-        ts.divergence(
-            sample_sets=[s for s in sample_sets.values()], 
-            indexes=list(itertools.combinations_with_replacement(range(len(strata)), 2)),
-            mode='branch', windows=inaccessible.position, span_normalise=True,
-        ) * (1 - filtered.rate)[:, np.newaxis] * mutation_rate
-    strata_divergence[inaccessible.rate == 0.0] = np.nan
+    strata_divergence = ts.divergence(
+        sample_sets=[s for s in sample_sets.values()], 
+        indexes=indexes,
+        mode='site',
+        windows=windows.position,
+        span_normalise=False,
+    )
+    strata_divergence[windows.rate == 1.0] /= \
+        accessible_bp[windows.rate == 1.0, np.newaxis]
+    strata_divergence[windows.rate == 0.0] = np.nan
     strata_stats["divergence"] = strata_divergence
 
-    strata_folded_afs = []
+    strata_afs = []
     for strata, sample_set in sample_sets.items():
-        folded_afs = ts.allele_frequency_spectrum(
-            sample_sets=[sample_set], mode='branch', 
-            windows=inaccessible.position, span_normalise=True,
-        )
-        folded_afs = np.sum(folded_afs * correction[:, np.newaxis] * mutation_rate, axis=0)
-        strata_folded_afs.append(folded_afs)
-    strata_stats["folded_afs"] = strata_folded_afs
+        afs = ts.allele_frequency_spectrum(
+            sample_sets=[sample_set], 
+            mode='site', 
+            polarised=snakemake.params.polarised,
+            span_normalise=False,
+        ) / accessible_bp[windows.rate == 1.0].sum()
+        strata_afs.append(afs)
+    strata_stats["afs"] = strata_afs
 
-    strata_unfolded_afs = []
-    for strata, sample_set in sample_sets.items():
-        unfolded_afs = ts.allele_frequency_spectrum(
-            sample_sets=[sample_set], mode='branch', polarised=True,
-            windows=inaccessible.position, span_normalise=True,
-        )
-        unfolded_afs = np.sum(unfolded_afs * correction[:, np.newaxis] * mutation_rate, axis=0)
-        strata_unfolded_afs.append(unfolded_afs)
-    strata_stats["unfolded_afs"] = strata_unfolded_afs
 pickle.dump(strata_stats, open(snakemake.output.strata_stats, "wb"))